python_implementation/main_computer_listeners/requirements.txt
paho-mqtt
numpy
opencv-python
# plyer # Optional: for desktop notifications in alert_listener.py 